Refactor the PR a smidge to make the change laser pointed to a new view - NodeTitleView
